relocate UITextfield extensions in separate file
