+++ iOS1-HW17-MaxMakeevV2/ViewController.swift
@@ -323,26 +323,3 @@ class ViewController: UIViewController {
         return stackView
     }
 }
-
-//MARK: - Extension to Textfield
-
-extension UITextField {
-    func setLeftIcon(_ image: UIImage) {
-        let iconView = UIImageView(frame: CGRect(x: 20, y: 5, width: 20, height: 20))
-        iconView.image = image
-        let iconContainerView: UIView = UIView(frame: CGRect(x: 20, y: 0, width: 60, height: 30))
-        iconContainerView.addSubview(iconView)
-        leftView = iconContainerView
-        leftViewMode = .always
-    }
-    
-    func setRightIcon(_ image: UIImage) {
-        let iconView = UIImageView(frame: CGRect(x: -10, y: 5, width: 20, height: 20))
-        iconView.image = image
-        let iconContainerView: UIView = UIView(frame: CGRect(x: 100, y: 0, width: 30, height: 30))
-        iconContainerView.addSubview(iconView)
-        rightView = iconContainerView
-        rightViewMode = .always
-    }
-}
-
